Split preprocessing into several modules, updated tests and docs accordingly
