+++ docs/function-docs/algorithm-tools.rst
@@ -17,26 +17,15 @@ Algorithm tools
 
 .. _wrapper-function-docs:
 
-vantage6.algorithm.tools.preprocessing
-++++++++++++++++++++++
-
-.. automodule:: vantage6.algorithm.tools.preprocessing
-   :members:
-   :show-inheritance:
-
-.. automodule:: vantage6.algorithm.tools.preprocessing.functions
-   :members:
-   :show-inheritance:
-
 vantage6.algorithm.tools.wrappers
-++++++++++++++++++++++
++++++++++++++++++++++++++++++++++
 
 .. automodule:: vantage6.algorithm.tools.wrappers
    :members:
    :show-inheritance:
 
 vantage6.algorithm.tools.wrap
-++++++++++++++++++++++
++++++++++++++++++++++++++++++
 
 .. automodule:: vantage6.algorithm.tools.wrap
    :members:
@@ -45,14 +34,14 @@ vantage6.algorithm.tools.wrap
 .. _mock-client-api-ref:
 
 vantage6.algorithm.tools.mock_client
-++++++++++++++++++++++++++
+++++++++++++++++++++++++++++++++++++
 
 .. automodule:: vantage6.algorithm.tools.mock_client
    :members:
 
 
 vantage6.algorithm.tools.util
-+++++++++++++++++++
++++++++++++++++++++++++++++++
 
 .. automodule:: vantage6.algorithm.tools.util
    :members:
