+++ docs/function-docs/preprocessing.rst
@@ -1,0 +1,52 @@
+Preprocessing
+=============
+
+Aggregation operations
+----------------------
+
+vantage6.algorithm.tools.preprocessing.aggregation
+++++++++++++++++++++++++++++++++++++++++++++++++++
+
+.. automodule:: vantage6.algorithm.tools.preprocessing.aggregation
+   :members:
+   :show-inheritance:
+
+Column operations
+----------------------
+
+vantage6.algorithm.tools.preprocessing.column
++++++++++++++++++++++++++++++++++++++++++++++
+
+.. automodule:: vantage6.algorithm.tools.preprocessing.column
+   :members:
+   :show-inheritance:
+
+Datetime operations
+----------------------
+
+vantage6.algorithm.tools.preprocessing.datetime
++++++++++++++++++++++++++++++++++++++++++++++++
+
+.. automodule:: vantage6.algorithm.tools.preprocessing.datetime
+   :members:
+   :show-inheritance:
+
+Encoding operations
+----------------------
+
+vantage6.algorithm.tools.preprocessing.encoding
++++++++++++++++++++++++++++++++++++++++++++++++
+
+.. automodule:: vantage6.algorithm.tools.preprocessing.encoding
+   :members:
+   :show-inheritance:
+
+Filter operations
+----------------------
+
+vantage6.algorithm.tools.preprocessing.filter
++++++++++++++++++++++++++++++++++++++++++++++
+
+.. automodule:: vantage6.algorithm.tools.preprocessing.filter
+   :members:
+   :show-inheritance:
