+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.black]
-line-length = 80
+line-length = 79
 
 include = '\.pyi?$'
 exclude = '''
