+++ vantage6-algorithm-tools/vantage6/algorithm/tools/preprocessing/__init__.py
@@ -3,8 +3,67 @@ import sys
 
 import pandas as pd
 
-import vantage6.algorithm.tools.preprocessing.functions as prepro_functions
+from vantage6.algorithm.tools.preprocessing.aggregation import (
+    collapse,
+    group_statistics,
+)
+from vantage6.algorithm.tools.preprocessing.column import (
+    assign_column,
+    rename_columns,
+)
+from vantage6.algorithm.tools.preprocessing.datetime import (
+    calculate_age,
+    timedelta,
+    to_datetime,
+    to_timedelta,
+)
+from vantage6.algorithm.tools.preprocessing.encoding import (
+    discretize_column,
+    encode,
+    extract_from_string,
+    impute,
+    min_max_scale,
+    one_hot_encode,
+    standard_scale,
+)
+from vantage6.algorithm.tools.preprocessing.filter import (
+    drop_columns,
+    drop_columns_by_index,
+    filter_by_date,
+    filter_range,
+    select_columns,
+    select_columns_by_index,
+    select_rows,
+)
 from vantage6.algorithm.tools.util import error
+
+funcs = [
+    assign_column,
+    calculate_age,
+    collapse,
+    discretize_column,
+    drop_columns,
+    drop_columns_by_index,
+    encode,
+    extract_from_string,
+    filter_by_date,
+    filter_range,
+    group_statistics,
+    impute,
+    min_max_scale,
+    one_hot_encode,
+    rename_columns,
+    select_columns,
+    select_columns_by_index,
+    select_rows,
+    standard_scale,
+    timedelta,
+    to_datetime,
+    to_timedelta,
+]
+func_dict = {f.__name__: f for f in funcs}
+
+__all__ = list(func_dict.keys())
 
 
 def preprocess_data(
@@ -40,9 +99,9 @@ def preprocess_data(
         func_name = preprocess_step["function"]
 
         # get preprocessing function
-        try:
-            func = getattr(prepro_functions, func_name)
-        except AttributeError:
+        if func_name in func_dict:
+            func = func_dict.get(func_name)
+        else:
             error(
                 f"Unknown preprocessing type '{func_name}' defined. Please "
                 "check your preprocessing input. Exiting..."
